Fixing compiler warnings (MSVC) for truncation of types

+++ Source/EbsdLib/Core/OrientationMath.cpp
@@ -82,7 +82,7 @@ void OrientationMath::RootTensorFromLatticeParameters(float a, float b, float c,
 // -----------------------------------------------------------------------------
 //
 // -----------------------------------------------------------------------------
-void OrientationMath::MillerBravaisToMillerDirection(const int millerBravais[4], int miller[3])
+void OrientationMath::MillerBravaisToMillerDirection(const int32_t millerBravais[4], int32_t miller[3])
 {
   miller[0] = millerBravais[0] - millerBravais[2];
   miller[1] = millerBravais[1] - millerBravais[2];
@@ -91,17 +91,17 @@ void OrientationMath::MillerBravaisToMillerDirection(const int millerBravais[4],
 // -----------------------------------------------------------------------------
 //
 // -----------------------------------------------------------------------------
-void OrientationMath::MillerToMillerBravaisDirection(const int miller[3], int millerBravais[4])
+void OrientationMath::MillerToMillerBravaisDirection(const int32_t miller[3], int32_t millerBravais[4])
 {
-  millerBravais[0] = 0.33333f * (2 * miller[0] - miller[1]);
-  millerBravais[1] = 0.33333f * (2 * miller[1] - miller[0]);
-  millerBravais[2] = -0.33333f * (miller[0] + miller[1]);
+  millerBravais[0] = static_cast<int32_t>(0.33333f * (2 * miller[0] - miller[1]));
+  millerBravais[1] = static_cast<int32_t>(0.33333f * (2 * miller[1] - miller[0]));
+  millerBravais[2] = static_cast<int32_t>(-0.33333f * (miller[0] + miller[1]));
   millerBravais[3] = miller[2];
 }
 // -----------------------------------------------------------------------------
 //
 // -----------------------------------------------------------------------------
-void OrientationMath::MillerBravaisToMillerPlane(const int millerBravais[4], int miller[3])
+void OrientationMath::MillerBravaisToMillerPlane(const int32_t millerBravais[4], int32_t miller[3])
 {
   miller[0] = millerBravais[0];
   miller[1] = millerBravais[1];
@@ -110,7 +110,7 @@ void OrientationMath::MillerBravaisToMillerPlane(const int millerBravais[4], int
 // -----------------------------------------------------------------------------
 //
 // -----------------------------------------------------------------------------
-void OrientationMath::MillerToMillerBravaisPlane(const int miller[3], int millerBravais[4])
+void OrientationMath::MillerToMillerBravaisPlane(const int32_t miller[3], int32_t millerBravais[4])
 {
   millerBravais[0] = miller[0];
   millerBravais[1] = miller[1];
